requirements.txt
tensorflow==2.7.0
pillow==8.2.0
streamlit==0.82.0